% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwtbl_add_rise_vars.R
\name{add_rise_vars_to_rwtbl}
\alias{add_rise_vars_to_rwtbl}
\title{Add required RISE variables to the rwtbl}
\usage{
add_rise_vars_to_rwtbl(rwtbl, ui_vars)
}
\arguments{
\item{rwtbl}{A tbl_df with expected columns. Likely this is output from
\code{\link[RWDataPlyr:rdf_to_rwtbl2]{RWDataPlyr::rdf_to_rwtbl2()}}}

\item{ui_vars}{User specified values (as a list) for the \code{sourceCode},
\code{modelNameSourceCode}, \code{status}, and \code{modelRunDescription} fields in the
RISE json file.}
}
\value{
tbl with all required fields for RISE json file
}
\description{
\code{rwtbl_add_rise_vars()} takes in a rwtbl and adds in the variables required
for publication in RISE.
}
