% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_variables.R
\docType{data}
\name{rwtbl_cols_for_rise}
\alias{rwtbl_cols_for_rise}
\alias{rise_json_req_obj}
\title{Required columns, variables, and objects}
\format{An object of class \code{character} of length 9.}
\usage{
rwtbl_cols_for_rise

rise_json_req_obj
}
\description{
\code{rwtbl_cols_for_rise()} contains the columns that are required from the
rdf file to ensure rdf2rise can fill in all required objects for RISE. This
should be passed to \code{\link[RWDataPlyr:rdf_to_rwtbl2]{RWDataPlyr::rdf_to_rwtbl2()}} as the \code{keep_cols}
argument. \code{rise_json_req_obj()} contains all of the required json objects for
RISE. All of these will be returned by \code{\link[=rwtbl_add_rise_vars]{rwtbl_add_rise_vars()}}, but this
could be used if you are creating your own tbl to pass to
\code{\link[=tbl_to_rise_json]{tbl_to_rise_json()}}.
}
\keyword{datasets}
