% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rise_json.R
\name{write_rise_json}
\alias{write_rise_json}
\title{Write the RISE json object to disk}
\usage{
write_rise_json(x, path)
}
\arguments{
\item{x}{json character vector. Likely output from \code{\link[=tbl_to_rise_json]{tbl_to_rise_json()}}.}

\item{path}{File path to save the file to. Should be an existing directory.}
}
\value{
Invisibly returns \code{x}. Saves \code{x} to \code{path} as .json file.
}
\description{
\code{write_rise_json()} takes the json character object and writes the data to
the specified location. RISE requires a file format named
sourceCode_YYYYMMDDHHMMSSMS.json, so this function creates a file following
that naming convention. \code{sourceCode} is a field that exists in the json
character object, so it is used for the file name.
}
