% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf2rise.R
\docType{package}
\name{rdf2rise-package}
\alias{rdf2rise}
\alias{rdf2rise-package}
\title{rdf2rise: Translate RDF Files to RISE JSON Files}
\description{
rdf2rise is a package to help translate 
  'RiverWare'(TM)<http://www.riverware.org/> "rdf"" files into the json files 
  necessary to publish data to the Reclamation Information Sharing Environment 
  (RISE). It is intended for pushing modeled data that is not stored in other 
  Reclamation databases to the RISE data base. It works by making guesses about 
  certain fields, so it can be used in a fully automatic fashion, but allows the 
  user to easily override those gueses in a tidy process.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BoulderCodeHub/rdf2rise}
  \item Report bugs at \url{https://github.com/BoulderCodeHub/rdf2rise/issues}
}

}
\author{
\strong{Maintainer}: Alan Butler \email{rabutler@usbr.gov}

}
\keyword{internal}
