% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_to_rise_json.R
\name{tbl_to_rise_json}
\alias{tbl_to_rise_json}
\title{Converts a tbl_df to the RISE JSON format}
\usage{
tbl_to_rise_json(tbl)
}
\arguments{
\item{tbl}{A tbl_df with all required columns (fields for the RISE json
file). Likely output from \code{\link[=rwtbl_add_rise_vars]{rwtbl_add_rise_vars()}}.}
}
\value{
Invisibly returns data in RISE json format
}
\description{
\code{tbl_to_rise_json()} converts a tbl to the RISE json format.
}
\details{
The RISE json format is a specific json format that (1) has one complete
object per row and (2) does not include any arrays. Each object on a row
contains 16 required object-value pairs.
}
