% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rise_json.R
\name{write_rise_json}
\alias{write_rise_json}
\title{Write the RISE json object to disk}
\usage{
write_rise_json(x, path)
}
\arguments{
\item{x}{json character vector. Likely output from \code{\link[=tbl_to_rise_json]{tbl_to_rise_json()}}.}

\item{path}{File path to save the file to. Should be an existing directory.}
}
\value{
Invisibly returns \code{x}. Saves \code{x} to \code{path} as .json file.
}
\description{
\code{write_rise_json()} takes the json character object and writes the data to
the specified location. RISE requires a file format named
sourceCode_YYYYMMDDHHMMSSMS.json, so this function creates a file following
that naming convention. \code{sourceCode} is a field that exists in the json
character object, so it is used for the file name.
}
\examples{
ifile <- system.file(
  'extdata/Scenario/ISM1988_2014,2007Dems,IG,Most',
  "KeySlots.rdf",
 package = "RWDataPlyr"
)

# get the tbl using RWDataPlyr
rwtbl <- RWDataPlyr::rdf_to_rwtbl2(
  ifile,
  scenario = "test",
  keep_cols = rwtbl_cols_for_rise
)

# manually specify some parameters:
ui_vars <- list(
  sourceCode = "CRSS-TestData",
  modelNameSourceCode = "CRSS",
  status = "Finalized Dec. 2012. To RISE v0.0.1",
  modelRunDescription = "desc"
)

# get the rest of the parameters automatically
rise_tbl <- rwtbl_add_rise_vars(rwtbl, ui_vars)

# add then convert to the RISE json format
rise_json <- tbl_to_rise_json(rise_tbl)

# finally write the file
write_rise_json(rise_json, path = tempdir())

}
